stream.dash: move DASHSegment to dedicated module
