+++ src/streamlink/stream/dash/__init__.py
@@ -1,2 +1,3 @@
 from streamlink.stream.dash.dash import DASHStream, DASHStreamReader, DASHStreamWorker, DASHStreamWriter
-from streamlink.stream.dash.manifest import MPD, DASHSegment, MPDParsingError
+from streamlink.stream.dash.manifest import MPD, MPDParsingError
+from streamlink.stream.dash.segment import DASHSegment
